agents/currency-agent/pyproject.toml
[project]
name = "currency-agent"
version = "1.0.0"
description = "Currency conversion agent with MCP Gateway support for OPA policy enforcement"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
server = "currency_agent.agent:a2a_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
